+++ gui/src/app/pages/NwbPage/viewPlugins/ElectricalSeriesItemView/SpikeSortingView/SpikeSortingViewOld.tsx
@@ -1,14 +1,14 @@
 import { RemoteH5FileX } from "@remote-h5-file/index";
 import {
-  PairioJob,
-  PairioJobDefinition,
-  PairioJobRequiredResources,
-} from "app/pairio/types";
+  DendroJob,
+  DendroJobDefinition,
+  DendroJobRequiredResources,
+} from "app/dendro/dendro-types";
 import useRoute from "app/useRoute";
 import { FunctionComponent, useMemo } from "react";
 import { useNwbFile } from "../../../NwbFileContext";
-import { getJobOutputUrl, removeLeadingSlash } from "../../CEBRA/PairioHelpers";
-import PairioItemView from "../../CEBRA/PairioItemView";
+import { getJobOutputUrl, removeLeadingSlash } from "../../CEBRA/DendroHelpers";
+import DendroItemView from "../../CEBRA/DendroItemView";
 import useTimeSeriesInfo from "../../TimeSeries/useTimeseriesInfo";
 
 type SpikeSortingViewProps = {
@@ -25,7 +25,7 @@ const getJobDefinition = (
   adjustableParameterValues: { [key: string]: any },
   inputFileUrl: string,
   path: string,
-): PairioJobDefinition => {
+): DendroJobDefinition => {
   return {
     appName,
     processorName,
@@ -65,7 +65,7 @@ const getJobDefinition = (
 
 const getRequiredResources = (
   _requireGpu: boolean,
-): PairioJobRequiredResources => {
+): DendroJobRequiredResources => {
   return {
     numCpus: 2,
     numGpus: 0,
@@ -120,7 +120,7 @@ const SpikeSortingView: FunctionComponent<SpikeSortingViewProps> = ({
   }, []);
 
   const jobFilter = useMemo(
-    () => (job: PairioJob) => {
+    () => (job: DendroJob) => {
       // make sure electrical_series_path matches
       const p = job.jobDefinition.parameters.find(
         (p) => p.name === "electrical_series_path",
@@ -133,7 +133,7 @@ const SpikeSortingView: FunctionComponent<SpikeSortingViewProps> = ({
   );
 
   const sortCandidateJobs = useMemo(
-    () => (jobs: PairioJob[]) => {
+    () => (jobs: DendroJob[]) => {
       // favor jobs that have longest segment_duration_sec
       return [...jobs].sort((a, b) => {
         const p1 = a.jobDefinition.parameters.find(
@@ -180,7 +180,7 @@ const SpikeSortingView: FunctionComponent<SpikeSortingViewProps> = ({
   }
 
   return (
-    <PairioItemView
+    <DendroItemView
       width={width}
       height={height}
       nwbUrl={nwbUrl}
@@ -204,7 +204,7 @@ const SpikeSortingView: FunctionComponent<SpikeSortingViewProps> = ({
 };
 
 type SpikeSortingOutputComponentProps = {
-  job: PairioJob;
+  job: DendroJob;
   width: number;
   nwbFile: RemoteH5FileX;
 };